refactor: model version labels
