+++ webui/react/src/pages/ModelVersionDetails/ModelVersionHeader.tsx
@@ -1,25 +1,23 @@
 import { LeftOutlined } from '@ant-design/icons';
 import { Breadcrumb, Button, Dropdown, Menu, Modal, Space } from 'antd';
 import type { MenuProps } from 'antd';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 
 import InfoBox, { InfoRow } from 'components/InfoBox';
 import InlineEditor from 'components/InlineEditor';
 import Link from 'components/Link';
 import { relativeTimeRenderer } from 'components/Table';
-import TagList from 'components/TagList';
 import Avatar from 'components/UserAvatar';
 import { useStore } from 'contexts/Store';
 import useModalModelDownload from 'hooks/useModal/Model/useModalModelDownload';
 import useModalModelVersionDelete from 'hooks/useModal/Model/useModalModelVersionDelete';
+import ModelVersionTagList from 'pages/ModelVersionDetails/ModelVersionTagList';
 import { paths } from 'routes/utils';
-import { getModelVersionLabels } from 'services/api';
 import CopyButton from 'shared/components/CopyButton';
 import Icon from 'shared/components/Icon/Icon';
 import { formatDatetime } from 'shared/utils/datetime';
 import { copyToClipboard } from 'shared/utils/dom';
 import { ModelVersion } from 'types';
-import handleError from 'utils/error';
 import { getDisplayName } from 'utils/user';
 
 import css from './ModelVersionHeader.module.scss';
@@ -46,7 +44,6 @@ const ModelVersionHeader: React.FC<Props> = ({
   onSaveName,
 }: Props) => {
   const { users } = useStore();
-  const [ modelVersionTags, setModelVersionTags ] = useState<string[]>();
   const [ showUseInNotebook, setShowUseInNotebook ] = useState(false);
 
   const {
@@ -62,15 +59,6 @@ const ModelVersionHeader: React.FC<Props> = ({
   const handleDownloadModel = useCallback(() => {
     openModelDownload(modelVersion);
   }, [ modelVersion, openModelDownload ]);
-
-  const fetchModelVersionLabels = useCallback(async () => {
-    try {
-      const modelVersionLabels = await getModelVersionLabels({});
-      setModelVersionTags(modelVersionLabels);
-    } catch (e) {
-      handleError(e, { silent: true });
-    }
-  }, []);
 
   const infoRows: InfoRow[] = useMemo(() => {
     return [ {
@@ -102,13 +90,13 @@ const ModelVersionHeader: React.FC<Props> = ({
     },
     {
       content: (
-        <TagList
+        <ModelVersionTagList
           disabled={modelVersion.model.archived}
           ghost={false}
-          tagCandidates={modelVersionTags}
           tags={modelVersion.labels ?? []}
           onChange={onUpdateTags}
         />
+
       ),
       label: 'Tags',
     } ] as InfoRow[];
@@ -119,7 +107,6 @@ const ModelVersionHeader: React.FC<Props> = ({
     modelVersion.lastUpdatedTime,
     modelVersion.model.archived,
     modelVersion.userId,
-    modelVersionTags,
     onSaveDescription,
     onUpdateTags,
     users,
@@ -196,10 +183,6 @@ my_model.load_state_dict(ckpt['models_state_dict'][0])`);
 
     return <Menu className={css.overflow} items={menuItems} onClick={onItemClick} />;
   }, [ actions ]);
-
-  useEffect(() => {
-    fetchModelVersionLabels();
-  }, [ fetchModelVersionLabels ]);
 
   return (
     <header className={css.base}>
